% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anidata-package.R
\docType{package}
\name{anidata-package}
\alias{anidata}
\alias{anidata-package}
\title{anidata: My students' one-stop shop for some data they have to use in the class}
\description{
The package provides some secondary data-sets used for teaching purposes.
}
\author{
\strong{Maintainer}: Ani Ruhil \email{ruhil@ohio.edu}

}
\keyword{internal}
